requirements.txt
requests
python-dotenv
matplotlib
numpy